Modifying msbayes.pl to report parameter values.
I am adding an option ('-p') to msbayes.pl that causes it to report
all of the theta and tau parameter values for each prior sample. This
extra info is now passed to msbayes.pl from the msprior executable
via the "TAU_PSI_TBL" line. msbayes.pl now parses this extra info and,
if the '-p' option is specified, it will include columns of these
parameter values in the output prior file.