More updating of regress_cli.r.
More work on the CLI and output of regress_cli.r. The script now
outputs two different files: One containing the adjusted posterior
samples of continuous parameters, and the other containing a
summary of the posteriors of all parameters. The former is a
tab-delimited file, and the latter is a Python confObj-formatted
file.