More work on regress_cli.r.
Adding more options to CLI. Allowing user to specify column indices
for stats and parameters, in addition to using prefixes. This will
make things easier for wrapper code. These options required
some tweaks to the main function of the script to make sure the
all indices stayed current when the dummy column is removed from
the vector of parameter names.

Also, adding the output of the parameters analyzed to the summary
file. This helps diagnose problems when the same column is specified
as both a continuous and discrete parameter.