Overhauling theta priors in msprior.c.
Switching uniform priors to gamma priors. Both descendant thetas
are now independently drawn from the same gamma prior. The
ancestral theta is also drawn from the same prior and scaled
be a multiplier. Everything gets converted into values relative
to the mean of the two descendant thetas (`spTheta`) so that
everything is still in terms that msDQH can understand.