Implementing beta prior on bottleneck magnitude.
Placing beta prior on the bottleneck proportion in `msprior`. The
magnitude of the bottleneck, as the proportion of the current
theta, is now drawn from beta distribution for each descendant
population. If the user specifies a negative value for either of the
beta shape parameters, the proportions will be fixed to 1.0 (i.e.,
no bottleneck). This required changes to `msprior.c` and `msprior.h`.