Implementing prior on ancestral theta.
Implementing option to place gamma-distributed prior on theta of
the ancestral population. This entailed changing variable names
and logic in `msprior.c` and `msprior.h`.

This replaces the old option of simply specifying a multiplier
for scaling the ancestral theta. If the user specifies a negative
value for the shape or scale parameter of the gamma prior, the
descendant gamma-distributed theta prior is used.