Fixing bugs in `msprior.c`.
Correcting the allocation of `divModels`, which is the array
of partition arrays, to have the correct dimensions.

Replacing `memset` call on `PSIarray` with a manual setting of
all elements in `PSIarray` to zero. I am not sure why the use
of `memset` was causing problems, but this brute force approach
fixes the problem for now.