More work on `eureject`.
Adding write functions to `rejection.h` and `rejection.c` for
`d_array`, `s_array`, `sample`, and `sample_array` structs.

Updating `main` to either write the stat means and std deviations
to stdout if `num_retain` is < 1, or write the retained samples
otherwise.

Adding `-d` command-line option which controls whether or not
the Euclidean distance of each sample is reported in the first
column of the stdout.

Filling out the `help()` message and some misc clean up of
unnecessary code.