Updating include statement in `setup.c`.
There unfortunately are duplicate copies of both `hashtab.h` and
`hashtab.c` within the msBayes distribution. One set of copies
is in `src/hashtab` directory and the others are in `src`.

All `include` statements throughout the msbayes code use the
`src/hashtab.h` duplicate, EXCEPT for `setup.c` which was including
the `src/hashtab/hashtab.h`. This was causing a problem with
building from the distribution, because `src/hashtab/hashtab.h` was
not included. This commit changes the include in `setup.c` to
`src/hashtab.c` to be consistent with the rest of the code.

However, while this is a quick fix, it is NOT sustainable to have
duplicate copies of code in a distribution. If the goal is to
bundle `hashtab`, then all includes will need to be updated
accordingly, and the `src/hashtab`.c/h copies should be removed!