Updating `generate_int_partitions` function.
Updating `generate_int_partitions` function of
`partition_combinatorics`.c/h. Rather than the caller supplying
the number to partition, the total number of partitions of that
number, and a 2-dimensional array of ints for storing the
partitions, the caller now only supplies the number to
partition.

The function now initiates an `i_array_2d` instance for storing
all the partitions and returns a pointer to it.