Adding better GSL support to configure scripts.
`configure.ac` now fully supports a `--with-gsl` argument. When
`--with-gsl=yes` configure will look for GSL in several ways/places.
When `--with-gsl=no` it will only build GSL-independent ABACUS
tools (`eureject` and `intpart`). When a directory is specified,
configure will first look for `gsl-config`, and then resorts to
manually setting up basic GSL configurations based on the path
provided (if it exists).