Updating `obsSumStats.pl`.
The script was looking for sequence files listed in the config file
relative to the current working directory, rather than relative to the
config file. This commit changes this behavior. Now, if the file is
not found by name (i.e., the full path is not given in the config),
it looks for the file relative to the directory of the config file.