Updating `msprior.c`.
Adding support for uniform priors on divergence times (tau).  This is
done in a hacky way, whereby if the user specifies a zero or negative
number for either `tauShape` or `tauScale`, the absolute values of these
options are used to specify the upper and lower limits of a uniform
distribution on tau.