Updating `msprior`.
Using new functions in `probability` module of `abacus` to allow more
prior options in `msprior`. Specifically, uniform distributions on theta
parameters can now be specified using the same hacky method as for the
tau parameters (specifying one or both of the gamma shape and scale
parameters as zero or negative).