Updating `init_globals` function in `initvars`.
Previously, this function was case sensitive and did *not* report any
issues if any unrecognized config options were encountered. It just
skipped them and used the default settings (!!). The function is now
case insensitive and reports an error and exits if it finds any
unrecognized lines that are not comments. Also, the function now
continues to scan the config file until it finds the "BEGIN SAMPLE_TBL"
line. This ensures that it will find any invalid lines after it "runs
out" of expected options.