Updating `init_globals` in `initvars`.
Fixing problem with config files re-generated by `msbayes.pl`.
`msbayes.pl` sneaks in the `prngSeed` option into the config,
which was causing `msprior` to choke as this is option is not
expected/processed by `msprior`. Adding check for `prngSeed` in
`init_globals` to suppress the invalid option error and exit.