Updating `msbayes.pl` and `msprior.c`.
Overhauling the per-simulation output of `msprior.c` and the parsing of
that output by `msbayes.pl.`

Changing `msprior` to output the entire `taxonTauArray` rather than just
the unique tau values. This allows the order of the taus to mean
something when the sort index "0" is used with `msbayes.pl`.

Also, scraping the old behavior of extra output from `msbayes.pl` if
`numTauClasses` is constrained (i.e., not zero). `msbayes.pl` should
always be used with the `-p` flag which negates the need for this extra
output. I.e., when `-p` is used `msbayes.pl` reports the extra
information anyway.