Adding sorting options to `sumstatsvector`.
Updating `sumStatsVector.c` and `msStatsDQH.c` to allow more (no)
sorting options and updating the default sorting option.

The executable `sumstatsvector` will now accept 4 more sort index values
(8--11). These specify to group loci by taxon and to calculate moments
across the loci for each taxon, but without re-sorting the resulting
moments.

There is now an `allow_sorting` variable that gets set in
`sumStatsVector.c` according to the command line argument that sets the
sorting index variable `gSortPattern`. If the sort index is 0 or 8--11,
`allow_sorting` is set to 0, which causes all calls to `qsort` in
`msStatsDQH.c` to be skipped. Four is then subtracted from
`gSortPattern` if its value is 8--11, so that it will follow the same
grouping and moment-calculating behavior as the values 4--7, but with
any sorting turned off.

THE NEW DEFAULT SORT INDEX IS NOW 11! This behaves the same as the old
default 7 in terms of grouping by taxon, and taking means of summary
stats over loci, BUT it does NO SORTING over taxa.