Updating perl scripts that call `sumstatsvector`.
Updating all perl scripts that call `sumstatsvector` (dpp-msbayes.pl,
msbayes.pl, and obsSumStats.pl) to allow use of the new sorting index
options in `sumstatsvector` and to update the default value for this
option.

THE DEFAULT SORT INDEX FOR ALL THREE OF THESE PERL SCRIPTS IS NOW TO USE
SORT INDEX 11! This groups loci by taxon, takes the mean of the summary
statistics over loci for each taxon, and DOES NOT RESORT THE RESULTING
MEANS OVER TAXA. The old default did the same, but resorted the summary
stat means among the taxa by pi_between.