Updating `msprior.c`.
Adding a lower bound of "0.000000000001" to all theta parameters,
because the coalescent simulator `msDQH` crashes if theta parameters are
zero.

Also, I am updating the hack where there was a check for small (i.e.,
0.0001) divergence times. In such simulations, the div time was set to
this arbitrary lower bound, and the bottleneck time was set to 0.5 of
this.  I am guessing this was to prevent unrealistic (and numerically
unstable?) changes in pop size. However, 0.0001 can be thousands of
generations which is not trivial. Also, rather than this weird hack of
the bottleneck time, it seems much better to simply have no bottleneck
if the div time is essentially zero. Accordingly, I lowered the
threshold and simply "turn off" the bottleneck if the time is below it
(I no longer adjust the div time or bottleneck time).