Adding ``timeInSubsPerSite`` option.
Adding option to scale time in expected substitutions per site.
Updating code to accommodate a ``timeInSubsPerSite`` option in the
config file which defaults to 0 (false), but if not zero (true), causes
time to be in expected subs per site. This required changes to:

    dpp-msbayes.pl
    msprior.c
    msprior.h
    setup.c

Most of the changes were required to handle the new option. The changing
of the time scale was as simple as setting ``meanTheta`` to 1.0.