Fixing errors in ``old-model``.
Fixing errors in ``old-model`` that were introduced with the addition of
the ``timeInSubsPerSite`` option. I forgot to declare the variables for
the new option in ``old-model/msprior.h``. These are now added. I also
deleted a curly brace on line 382 when I added the new option to the
``SetDefaultParams`` option. This has been added back in.