Fixing rounding issues in perl scripts.
Fixing rounding issues in the calculation of the variance in divergence
times in the ``SummarizeTau`` function in the perl scripts
``dpp-msbayes.pl`` and ``old-model/msbayes.pl``. Due to rounding error,
the variance would often be negative, which caused an error when the
coefficient of variation was calculated (taking the square root of
negative number). There is now a check for negative variance values, in
which case the variance is set to zero.